+++ telegram/base.py
@@ -19,7 +19,7 @@
 """Base class for Telegram Objects."""
 
 try:
-    import ujson as json
+    import rapidjson as json
 except ImportError:
     import json