+++ tests/test_invoice.py
@@ -113,7 +113,7 @@ class TestInvoice(object):
     def test_send_object_as_provider_data(self, monkeypatch, bot, chat_id, provider_token):
         def test(_, url, data, **kwargs):
             return (data['provider_data'] == '{"test_data": 123456789}' or  # Depends if using
-                    data['provider_data'] == '{"test_data":123456789}')  # ujson or not
+                    data['provider_data'] == '{"test_data":123456789}')  # rapidjson or not
 
         monkeypatch.setattr('telegram.utils.request.Request.post', test)